Rename 'load_rules' module as (private) '_grib1_load_rules'.
