+++ iris_grib/tests/unit/grib1_load_rules/__init__.py
@@ -14,7 +14,7 @@
 #
 # You should have received a copy of the GNU Lesser General Public License
 # along with iris-grib.  If not, see <http://www.gnu.org/licenses/>.
-"""Unit tests for the :mod:`iris_grib.load_rules` module."""
+"""Unit tests for the :mod:`iris_grib._grib1_load_rules` module."""
 
 from __future__ import (absolute_import, division, print_function)
 from six.moves import (filter, input, map, range, zip)  # noqa
